Add projects using the new dotnet sdk
These projects are built using the dotnet tooling instead of the
current system MSBuild has. Long term, MSBuild wants to start using
the SDK as well (see
https://github.com/Microsoft/msbuild/issues/1399), so we'll get ahead
of the game for the projects we care about for the CLI, which makes
building using just dotnet easier (no more project.json's to worry
about, for example).